3_Model_Serving/requirements.txt
# FastAPI and web server dependencies
fastapi
uvicorn[standard]
pydantic

# Data processing
pandas
numpy

# Database
psycopg2-binary

# Model serving
scikit-learn
xgboost

# Model serialization
joblib

# Experiment tracking
mlflow

# Testing
locust